+++ src/creatures/combat/spells.h
@@ -41,13 +41,13 @@ class Spells final : public Scripts {
 		}
 
 		Spell* getSpellByName(const std::string &name);
-		RuneSpell* getRuneSpell(uint32_t id);
+		RuneSpell* getRuneSpell(uint16_t id);
 		RuneSpell* getRuneSpellByName(const std::string &name);
 
 		InstantSpell* getInstantSpell(const std::string &words);
 		InstantSpell* getInstantSpellByName(const std::string &name);
 
-		InstantSpell* getInstantSpellById(uint32_t spellId);
+		InstantSpell* getInstantSpellById(uint16_t spellId);
 
 		TalkActionResult_t playerSaySpell(Player* player, std::string &words);
 
@@ -130,10 +130,10 @@ class Spell : public BaseSpell {
 		void setName(std::string n) {
 			name = n;
 		}
-		uint8_t getId() const {
+		uint16_t getId() const {
 			return spellId;
 		}
-		void setId(uint8_t id) {
+		void setId(uint16_t id) {
 			spellId = id;
 		}
 
@@ -306,7 +306,7 @@ class Spell : public BaseSpell {
 		uint32_t magLevel = 0;
 		int32_t range = -1;
 
-		uint8_t spellId = 0;
+		uint16_t spellId = 0;
 
 		bool selfTarget = false;
 		bool needTarget = false;
